20200805 algorithm commit boj - q2562, q2577, q3052, q10818, q1244 swea - q1225
