+++ src/com/boj/bronze/q11022/Main.java
@@ -1,0 +1,31 @@
+package com.boj.bronze.q11022;
+
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.OutputStreamWriter;
+import java.util.StringTokenizer;
+
+public class Main {
+    public static void main(String[] args) throws NumberFormatException, IOException {
+
+        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+        BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
+
+        int T = Integer.parseInt(br.readLine());
+
+        for (int tc = 1; tc <= T; tc++) {
+            String input = br.readLine();
+            StringTokenizer st = new StringTokenizer(input);
+            int A = Integer.parseInt(st.nextToken());
+            int B = Integer.parseInt(st.nextToken());
+
+            bw.write("Case #" + tc + ": " + A + " + " + B + " = " + (A + B) + "\n");
+        }
+
+        bw.flush();
+        br.close();
+        bw.close();
+    }
+}
