+++ src/com/boj/bronze/q15552/Main.java
@@ -1,0 +1,26 @@
+package com.boj.bronze.q15552;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.util.StringTokenizer;
+
+public class Main {
+    public static void main(String[] args) throws NumberFormatException, IOException {
+
+        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+
+        int T = Integer.parseInt(br.readLine());
+
+        for(int tc = 0; tc < T; tc++) {
+            String line = br.readLine();
+            StringTokenizer st = new StringTokenizer(line);
+
+            int A = Integer.parseInt(st.nextToken());
+            int B = Integer.parseInt(st.nextToken());
+
+            System.out.println(A+B);
+        }
+    }
+}
+
