+++ src/com/boj/bronze/q9498/Main.java
@@ -1,0 +1,31 @@
+package com.boj.bronze.q9498;
+
+import java.util.Scanner;
+
+/**
+ *   시험 성적 계산
+ */
+public class Main {
+    public static void main(String[] args) {
+
+        Scanner sc = new Scanner(System.in);
+
+        int score = sc.nextInt();
+        char grade;
+
+        if (score >= 90 && score <= 100)
+            grade = 'A';
+        else if (score >= 80 && score < 90)
+            grade = 'B';
+        else if (score >= 70 && score < 80)
+            grade = 'C';
+        else if (score >= 60 && score < 70)
+            grade = 'D';
+        else
+            grade = 'F';
+
+        System.out.println(grade);
+
+        sc.close();
+    }
+}
