--- src/com/boj/q10950/Main.java
@@ -1,25 +1,0 @@
-package com.boj.q10950;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.StringTokenizer;
-
-public class Main {
-    public static void main(String[] args) throws NumberFormatException, IOException {
-
-        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-
-        int T = Integer.parseInt(br.readLine());
-
-        for(int tc = 0; tc < T; tc ++) {
-            String line = br.readLine();
-            StringTokenizer st = new StringTokenizer(line);
-
-            int A = Integer.parseInt(st.nextToken());
-            int B = Integer.parseInt(st.nextToken());
-
-            System.out.println(A+B);
-        }
-    }
-}
