--- src/com/boj/q10951/Main.java
@@ -1,19 +1,0 @@
-package com.boj.q10951;
-
-import java.io.*;
-import java.util.StringTokenizer;
-
-public class Main {
-    public static void main(String[] args) throws IOException {
-
-        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        String input = "";
-        while((input = br.readLine()) != null){
-            StringTokenizer st = new StringTokenizer(input);
-            int A = Integer.parseInt(st.nextToken());
-            int B = Integer.parseInt(st.nextToken());
-            System.out.println(A+B);
-        }
-        br.close();
-    }
-}
