--- src/com/boj/q10952/Main.java
@@ -1,28 +1,0 @@
-package com.boj.q10952;
-
-import java.io.*;
-import java.util.StringTokenizer;
-
-public class Main {
-    public static void main(String[] args) throws IOException {
-
-        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
-        String input = br.readLine();
-        StringTokenizer st = new StringTokenizer(input);
-
-        int A = Integer.parseInt(st.nextToken());
-        int B = Integer.parseInt(st.nextToken());
-
-        while (A != 0 || B != 0) {
-            bw.write("" + (A + B) + "\n");
-            input = br.readLine();
-            st = new StringTokenizer(input);
-            A = Integer.parseInt(st.nextToken());
-            B = Integer.parseInt(st.nextToken());
-        }
-        bw.flush();
-        br.close();
-        bw.close();
-    }
-}
