--- src/com/boj/q2739/Main.java
@@ -1,18 +1,0 @@
-package com.boj.q2739;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-
-public class Main {
-    public static void main(String[] args) throws NumberFormatException, IOException {
-
-        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-
-        int N = Integer.parseInt(br.readLine());
-
-        for (int i = 1; i < 10; i++) {
-            System.out.printf("%d * %d = %d\n", N, i, N * i);
-        }
-    }
-}
