--- src/com/boj/q2742/Main.java
@@ -1,25 +1,0 @@
-package com.boj.q2742;
-
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
-
-public class Main {
-    public static void main(String[] args) throws NumberFormatException, IOException {
-
-        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
-
-        int N = Integer.parseInt(br.readLine());
-
-        for (int i = N; i > 0; i--) {
-            bw.write("" + i + "\n");
-        }
-
-        bw.flush();
-        br.close();
-        bw.close();
-    }
-}
