--- src/com/boj/q8393/Main.java
@@ -1,21 +1,0 @@
-package com.boj.q8393;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-
-public class Main {
-    public static void main(String[] args) throws NumberFormatException, IOException {
-
-        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-
-        int n = Integer.parseInt(br.readLine());
-        int sum = 0;
-
-        for (int i = 1; i <= n; i++) {
-            sum += i;
-        }
-
-        System.out.println(sum);
-    }
-}
